REST api finishing touches

+++ backend/controllers/user.go
@@ -8,20 +8,13 @@ import (
 
 	"github.com/batijo/video-platform/backend/models"
 	"github.com/batijo/video-platform/backend/utils"
+	"github.com/batijo/video-platform/backend/utils/auth"
+	"github.com/jinzhu/gorm"
 
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/gorilla/mux"
 	"golang.org/x/crypto/bcrypt"
 )
-
-// ErrorResponse ...
-type ErrorResponse struct {
-	Err string
-}
-
-type error interface {
-	Error() string
-}
 
 // Login ...
 func Login(w http.ResponseWriter, r *http.Request) {
@@ -29,10 +22,12 @@ func Login(w http.ResponseWriter, r *http.Request) {
 	err := json.NewDecoder(r.Body).Decode(user)
 	if err != nil {
 		resp := models.Response{Status: false, Message: "Invalid request", Error: err.Error()}
+		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(resp)
 		return
 	}
 	resp := findOne(user.Email, user.Password)
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(resp)
 }
 
@@ -59,7 +54,6 @@ func findOne(email, password string) models.Response {
 
 	tk := &models.Token{
 		UserID: user.ID,
-		Name:   user.Name,
 		Email:  user.Email,
 		Admin:  user.Admin,
 		StandardClaims: &jwt.StandardClaims{
@@ -84,7 +78,13 @@ func findOne(email, password string) models.Response {
 func CreateUser(w http.ResponseWriter, r *http.Request) {
 
 	user := &models.User{}
-	json.NewDecoder(r.Body).Decode(user)
+	err := json.NewDecoder(r.Body).Decode(user)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Invalid request", Error: err.Error()}
+		w.WriteHeader(http.StatusBadRequest)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
 
 	// Check if user trying to gain admin access
 	if user.Admin {
@@ -122,32 +122,78 @@ func CreateUser(w http.ResponseWriter, r *http.Request) {
 
 // FetchUsers function
 func FetchUsers(w http.ResponseWriter, r *http.Request) {
-	var users []models.User
-	utils.DB.Preload("auths").Find(&users)
+	var (
+		users []models.User
+		res   *gorm.DB
+	)
+
+	userId, admin, err := auth.GetUserID(r)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Could not authorise user", Error: err.Error()}
+		w.WriteHeader(http.StatusUnauthorized)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	if admin {
+		res = utils.DB.Preload("auths").Find(&users)
+	} else {
+		res = utils.DB.Preload("auths").Where("ID = ? AND Public = ?", userId, true).Find(&users)
+	}
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "Could not fetch users", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
 
 	resp := models.Response{Status: true, Message: "Success", Data: users}
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(resp)
 }
 
 // UpdateUser ...
 func UpdateUser(w http.ResponseWriter, r *http.Request) {
+	var (
+		user models.User
+		id   = mux.Vars(r)["id"]
+	)
 
-	var user models.User
-
-	params := mux.Vars(r)
-	var id = params["id"]
-	res := utils.DB.First(&user, id)
-
-	if res.Error != nil {
-		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+	userId, admin, err := auth.GetUserID(r)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Could not authorise user", Error: err.Error()}
+		w.WriteHeader(http.StatusUnauthorized)
 		json.NewEncoder(w).Encode(resp)
 		return
 	}
-	json.NewDecoder(r.Body).Decode(&user)
+
+	res := utils.DB.First(&user, id)
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	if userId != user.ID || !admin {
+		resp := models.Response{Status: false, Message: "You have no privilage to perform this action"}
+		w.WriteHeader(http.StatusForbidden)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	err = json.NewDecoder(r.Body).Decode(&user)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Invalid request", Error: err.Error()}
+		w.WriteHeader(http.StatusBadRequest)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
 
 	// Check if user trying to gain admin access
 	if user.Admin {
 		resp := models.Response{Status: false, Message: "You can not make yourself an admin"}
+		w.WriteHeader(http.StatusForbidden)
 		json.NewEncoder(w).Encode(resp)
 		return
 	}
@@ -157,52 +203,104 @@ func UpdateUser(w http.ResponseWriter, r *http.Request) {
 		pass, err := bcrypt.GenerateFromPassword([]byte(user.Password), bcrypt.DefaultCost)
 		if err != nil {
 			resp := models.Response{Status: false, Message: "Password Encryption failed", Error: err.Error()}
+			w.WriteHeader(http.StatusInternalServerError)
 			json.NewEncoder(w).Encode(resp)
 			return
 		}
-
 		user.Password = string(pass)
 	}
 
-	utils.DB.Save(&user)
+	res = utils.DB.Save(&user)
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "Could not save user", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
 	resp := models.Response{Status: true, Message: "User updated", Data: user}
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(resp)
 }
 
 // DeleteUser ...
 func DeleteUser(w http.ResponseWriter, r *http.Request) {
-	params := mux.Vars(r)
-	var id = params["id"]
-	var user models.User
-	res := utils.DB.First(&user, id)
+	var (
+		id   = mux.Vars(r)["id"]
+		user models.User
+	)
 
-	// For some reason if you try to delete user which does not exist it deletes all users
-	if res.Error != nil {
-		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+	userId, admin, err := auth.GetUserID(r)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Could not authorise user", Error: err.Error()}
+		w.WriteHeader(http.StatusUnauthorized)
 		json.NewEncoder(w).Encode(resp)
 		return
 	}
 
-	utils.DB.Delete(&user)
+	res := utils.DB.First(&user, id)
+	// For some reason if you try to delete user which does not exist it deletes all users
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	if userId != user.ID || !admin {
+		resp := models.Response{Status: false, Message: "You have no privilage to perform this action"}
+		w.WriteHeader(http.StatusForbidden)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	res = utils.DB.Delete(&user)
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "Could not delete user", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusInternalServerError)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
 
 	resp := models.Response{Status: true, Message: "User deleted", Data: user}
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(resp)
 }
 
 // GetUser ...
 func GetUser(w http.ResponseWriter, r *http.Request) {
-	params := mux.Vars(r)
-	var id = params["id"]
-	var user models.User
-	res := utils.DB.Preload("Video").Preload("Video.AudioT").Preload("Video.SubtitleT").First(&user, id)
+	var (
+		id   = mux.Vars(r)["id"]
+		user models.User
+	)
 
-	if res.Error != nil {
-		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+	userId, admin, err := auth.GetUserID(r)
+	if err != nil {
+		resp := models.Response{Status: false, Message: "Could not authorise user", Error: err.Error()}
+		w.WriteHeader(http.StatusUnauthorized)
 		json.NewEncoder(w).Encode(resp)
 		return
 	}
 
+	res := utils.DB.Preload("Video").Preload("Video.AudioT").Preload("Video.SubtitleT").First(&user, id)
+	if res.Error != nil {
+		resp := models.Response{Status: false, Message: "User not found", Error: res.Error.Error()}
+		w.WriteHeader(http.StatusNotFound)
+		json.NewEncoder(w).Encode(resp)
+		return
+	}
+
+	if !user.Public || !admin {
+		if userId != user.ID {
+			resp := models.Response{Status: false, Message: "You have no privilage to perform this action"}
+			w.WriteHeader(http.StatusForbidden)
+			json.NewEncoder(w).Encode(resp)
+			return
+		}
+	}
+
 	resp := models.Response{Status: true, Message: "Success", Data: user}
+	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(resp)
 }
